+++ backend/tests/test_csv_exporters/helpers.py
@@ -62,6 +62,6 @@ class CSVWriterMock(mock.MagicMock):
 
 @contextmanager
 def mock_csv_writer() -> Generator[CSVWriterMock]:
-    path = "tilavarauspalvelu.services.csv_export._base_exporter.csv.writer"
+    path = "tilavarauspalvelu.services.export._base_exporter.csv.writer"
     with mock.patch(path, new_callable=CSVWriterMock) as mock_file:
         yield mock_file
